Refactor migration
