+++ CLAUDE.md
@@ -43,57 +43,17 @@ The plugin integrates deeply with:
 - Store `ap_actor_id` references instead of duplicating actor metadata in post meta
 - The `attributedTo` field should only be set for reblogs, not regular posts
 
-### Testing
-- Run tests with `composer test`
-- ActivityPub tests require the ActivityPub plugin to be available
+### Testing & Code Quality
+- `composer test` - Run tests (ActivityPub tests require the ActivityPub plugin)
+- `composer check-cs` - Check coding standards
+- `composer fix-cs` - Auto-fix coding standards issues
 
 ## WordPress Playground Testing
 
-When working on this WordPress plugin, you can help users test changes by providing a WordPress Playground link that uses the current branch.
-
-### How to Generate Playground Links
-
-When you've made changes to the plugin and pushed to a branch, provide a WordPress Playground link in this format:
+After pushing changes to a branch, provide a WordPress Playground link for testing:
 
 ```
 https://playground.wordpress.net/#{%22steps%22:[{%22step%22:%22installPlugin%22,%22pluginData%22:{%22resource%22:%22git:directory%22,%22url%22:%22https://github.com/akirk/friends%22,%22ref%22:%22BRANCH_NAME%22,%22refType%22:%22branch%22},%22options%22:{%22activate%22:true}}]}
 ```
 
-Replace `BRANCH_NAME` with the actual branch name you're working on.
-
-### Example
-
-If working on branch `claude/add-feature-xyz`, the link would be:
-
-```
-https://playground.wordpress.net/#{%22steps%22:[{%22step%22:%22installPlugin%22,%22pluginData%22:{%22resource%22:%22git:directory%22,%22url%22:%22https://github.com/akirk/friends%22,%22ref%22:%22claude/add-feature-xyz%22,%22refType%22:%22branch%22},%22options%22:{%22activate%22:true}}]}
-```
-
-### When to Provide Links
-
-- After completing a feature implementation
-- After fixing a bug
-- When the user requests to test changes
-- After creating or updating a pull request
-
-### Message Format
-
-When providing the link, use this format:
-
-```markdown
-## Test These Changes in WordPress Playground
-
-You can test the changes from branch `BRANCH_NAME` directly in WordPress Playground:
-
-[Launch WordPress Playground](PLAYGROUND_URL)
-
-This will install and activate the plugin with the changes from this branch.
-```
-
-### Notes
-
-- The playground link uses the GitHub repository as the source
-- The `installPlugin` step automatically installs and activates the plugin
-- Users can test the changes without needing to set up a local WordPress environment
-- This mirrors the functionality of the GitHub action in `.github/workflows/pr-playground-comment.yml` but uses branches instead of PR refs
-- Important: Use `"refType":"branch"` when referencing branch names, not `"refType":"refname"` (which is used for full refs like `refs/pull/123/head`)
+Replace `BRANCH_NAME` with the branch name. Use `"refType":"branch"` for branches (not `"refType":"refname"` which is for full refs like `refs/pull/123/head`).
